:recycle:  shuffle around the code with minor changes
