+++ packages/printer-legacy/src/printer.js
@@ -19,8 +19,7 @@ const {
 
 const { printRelations } = require("./relation");
 const { printDescription } = require("./common");
-const { printCustomDirectives } = require("./directive");
-const { printCustomTags } = require("./badge");
+const { printCustomDirectives, printCustomTags } = require("./directive");
 const {
   printOperationMetadata,
   printCodeOperation,
@@ -144,7 +143,7 @@ class Printer {
 
   static printCustomDirectives = printCustomDirectives;
 
-  static printBadges = printCustomTags;
+  static printCustomTags = printCustomTags;
 
   static printTypeMetadata = (type, options) => {
     let metadata;
@@ -205,14 +204,14 @@ class Printer {
       type,
       printTypeOptions,
     );
-    // const badges = Printer.printBadges(type, printTypeOptions);
+    const tags = Printer.printCustomTags(type, printTypeOptions);
     const metadata = Printer.printTypeMetadata(type, printTypeOptions);
     const relations = Printer.printRelations(type, printTypeOptions);
 
     return [
       header,
       mdx,
-      // badges,
+      tags,
       description,
       code,
       customDirectives,
